requirements.txt
pandas
numpy
vanna[chromadb,openai,postgres]
python-dotenv
